requirements.txt
flask
waitress
pillow